+++ samples/Basic/Pages/Index.cshtml.cs
@@ -57,7 +57,6 @@ namespace Basic.Pages
 				Users = await keysetContext.Query
 				  .Take(size)
 				  .ToListAsync();
-				Users.Reverse();
 			}
 			else if (after != null)
 			{
@@ -74,7 +73,6 @@ namespace Basic.Pages
 				Users = await keysetContext.Query
 				  .Take(size)
 				  .ToListAsync();
-				Users.Reverse();
 			}
 			else
 			{
@@ -83,6 +81,8 @@ namespace Basic.Pages
 				  .Take(size)
 				  .ToListAsync();
 			}
+
+			keysetContext.EnsureCorrectOrder(Users);
 
 			Elapsed = sw.ElapsedMilliseconds.ToString();